+++ src/App.jsx
@@ -1,18 +1,17 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import { updateTaskTitle, addTask, deleteTask } from './actions';
+import { updateTaskTitle, addTask } from './actions';
 
 import Page from './Page';
 
 function selector(state) {
   return {
-    tasks: state.tasks,
     taskTitle: state.taskTitle,
   };
 }
 
 export default function App() {
-  const { tasks, taskTitle } = useSelector(selector);
+  const { taskTitle } = useSelector(selector);
 
   const dispatch = useDispatch();
 
@@ -24,17 +23,11 @@ export default function App() {
     dispatch(addTask());
   }
 
-  function handleClickDeleteTask(id) {
-    dispatch(deleteTask(id));
-  }
-
   return (
     <Page
       taskTitle={taskTitle}
       onChangeTitle={handleChangeTitle}
       onClickAddTask={handleClickAddTask}
-      tasks={tasks}
-      onClickDeleteTask={handleClickDeleteTask}
     />
   );
 }
