+++ src/ListContainer.test.jsx
@@ -1,0 +1,21 @@
+import { render } from '@testing-library/react';
+
+import { useSelector } from 'react-redux';
+
+import ListContainer from './ListContainer';
+
+import tasks from '../fixtures/tasks';
+
+jest.mock('react-redux');
+
+test('ListContainer', () => {
+  useSelector.mockImplementation((selector) => selector({
+    tasks,
+  }));
+
+  const { getByText } = render((
+    <ListContainer />
+  ));
+
+  expect(getByText(/아무것도 하지 않기 #1/)).not.toBeNull();
+});
