+++ src/Page.test.jsx
@@ -1,29 +1,31 @@
 import { render, fireEvent } from '@testing-library/react';
 
+import { useSelector } from 'react-redux';
 import Page from './Page';
 
+import tasks from '../fixtures/tasks';
+
+jest.mock('react-redux');
+
 test('Page', () => {
+  useSelector.mockImplementation((selector) => selector({
+    tasks,
+    taskTitle: '',
+  }));
+
   const handleChangeTitle = jest.fn();
   const handleClickAddTask = jest.fn();
-  const handleClickDeleteTask = jest.fn();
-
-  const tasks = [
-    { id: 1, title: 'Task-1' },
-    { id: 2, title: 'Task-2' },
-  ];
 
   const { getByText } = render((
     <Page
       taskTitle=""
       onChangeTitle={handleChangeTitle}
       onClickAddTask={handleClickAddTask}
-      tasks={tasks}
-      onClickDeleteTask={handleClickDeleteTask}
     />
   ));
 
-  expect(getByText(/Task-1/)).not.toBeNull();
-  expect(getByText(/Task-2/)).not.toBeNull();
+  expect(getByText(/아무것도 하지 않기 #1/)).not.toBeNull();
+  expect(getByText(/아무것도 하지 않기 #2/)).not.toBeNull();
 
   fireEvent.click(getByText('추가'));
